vk-crawler/python-summarizer/requirements.txt
Flask
Flask-Cors
lexrank
razdel
grpcio
grpcio-tools
pyspark